+++ samples/interactive/int12-switchingTextureSets/hostCode.cpp
@@ -66,7 +66,7 @@ struct BoxAnimState {
   {
     vec3f rel = (vec3f(boxID)+.5f) / vec3f(numBoxes);
     boxCenter = vec3f(-worldSize) + (2.f*worldSize)*rel;
-    rotationAngle0 = distribution_uniform(rndGen)*(2.f*M_PI);
+    rotationAngle0 = float(distribution_uniform(rndGen)*(2.f*M_PI));
     do {
       rotationAxis.x = distribution_uniform(rndGen);
       rotationAxis.y = distribution_uniform(rndGen);
@@ -96,11 +96,11 @@ void addFace(Mesh &mesh,
              int faceID,
              const vec3f ll, const vec3f du, const vec3f dv)
 {
-  int idxll = mesh.vertices.size();
+  int idxll = (int)mesh.vertices.size();
   for (int iy=0;iy<2;iy++)
     for (int ix=0;ix<2;ix++) {
       mesh.vertices.push_back(ll+float(ix)*du+float(iy)*dv);
-      mesh.texCoords.push_back(vec2f(ix,iy));
+      mesh.texCoords.push_back(vec2f((float)ix,(float)iy));
     }
   mesh.indices.push_back(vec4i(idxll,idxll+1,idxll+3,faceID));
   mesh.indices.push_back(vec4i(idxll,idxll+3,idxll+2,faceID));
@@ -159,10 +159,10 @@ OWLGroup createBox(OWLContext context,
     // create a 4x4 checkerboard texture
     // ------------------------------------------------------------------
     vec2i texSize(distribution_texSize(rndGen),distribution_texSize(rndGen));
-    vec4uc color0 = vec4uc(255.99f*vec4f(distribution_uniform(rndGen),
-                                         distribution_uniform(rndGen),
-                                         distribution_uniform(rndGen),
-                                         0.f));
+    vec4uc color0 = vec4uc(255.99f*vec4f((float)distribution_uniform(rndGen),
+        (float)distribution_uniform(rndGen),
+        (float)distribution_uniform(rndGen),
+                             0.f));
     vec4uc color1 = vec4uc(255)-color0;
     std::vector<vec4uc> texels;
     for (int iy=0;iy<texSize.y;iy++)
@@ -261,7 +261,7 @@ void Viewer::cameraChanged()
   owlRayGenSet3f    (rayGen,"camera.dir_00",(const owl3f&)camera_d00);
   owlRayGenSet3f    (rayGen,"camera.dir_du",(const owl3f&)camera_ddu);
   owlRayGenSet3f    (rayGen,"camera.dir_dv",(const owl3f&)camera_ddv);
-  vec3f lightDir = (1,1,1);
+  vec3f lightDir = (1.f,1.f,1.f);
   owlRayGenSet3f    (rayGen,"lightDir",     (const owl3f&)lightDir);
   sbtDirty = true;
 }
@@ -309,7 +309,7 @@ Viewer::Viewer()
   for (int iz=0;iz<numBoxes.z;iz++)
     for (int iy=0;iy<numBoxes.y;iy++)
       for (int ix=0;ix<numBoxes.x;ix++) {
-        int ID = boxAnimStates.size();
+        int ID = (int)boxAnimStates.size();
         boxAnimStates.push_back(BoxAnimState());
         boxAnimStates[ID].init(vec3i(ix,iy,iz));
         boxTransforms.push_back(boxAnimStates[ID].getTransform(0.f));
@@ -403,7 +403,7 @@ void Viewer::render()
   static double t0 = getCurrentTime();
   double t = animSpeed * (getCurrentTime() - t0);
   for (int i=0;i<boxTransforms.size();i++) {
-    boxTransforms[i] = boxAnimStates[i].getTransform(t);
+    boxTransforms[i] = boxAnimStates[i].getTransform((float)t);
     owlInstanceGroupSetTransform(world,i,
                                  (const float*)&boxTransforms[i],
                                  OWL_MATRIX_FORMAT_OWL);
@@ -421,7 +421,7 @@ void Viewer::render()
   // owlGroupBuildAccel(world);
   owlBuildSBT(context);
 
-  float t_rel = getCurrentTime() - t0;
+  float t_rel = float(getCurrentTime() - t0);
   owlParamsSet1f(lp,"time",t_rel);
   owlParamsSet1i(lp,"deviceCount",owlGetDeviceCount(context));
   owlLaunch2D(rayGen,fbSize.x,fbSize.y,lp);
